# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reward Loyalty API",
        "description": "API for the Reward Loyalty script.",
        "contact": {
            "email": "support@nowsquare.com"
        },
        "version": "1.3.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        },
        {
            "url": "https://reward-loyalty-demo.nowsquare.com/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/{locale}/v1/member/register": {
            "post": {
                "tags": [
                    "Member",
                    "Authentication"
                ],
                "summary": "Register a new member",
                "description": "This endpoint allows a new member to register an account.",
                "operationId": "registerMember",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "The locale (e.g., `en-us`)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "en-us"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The email of the member",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "email",
                            "example": "email@example.com"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "The name of the member",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 64,
                            "example": "John Doe"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The password of the member, if empty a password is generated",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 48,
                            "minLength": 6,
                            "example": "password123"
                        }
                    },
                    {
                        "name": "time_zone",
                        "in": "query",
                        "description": "The time zone of the member",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "America/New_York"
                        }
                    },
                    {
                        "name": "locale",
                        "in": "query",
                        "description": "The locale of the member (e.g., `en_US`)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 12,
                            "minLength": 5,
                            "example": "en_US"
                        }
                    },
                    {
                        "name": "currency",
                        "in": "query",
                        "description": "The currency of the member (e.g., `USD`)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 3,
                            "minLength": 3,
                            "example": "USD"
                        }
                    },
                    {
                        "name": "accepts_emails",
                        "in": "query",
                        "description": "Whether the member accepts marketing emails from a partner or not",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ],
                            "example": 0
                        }
                    },
                    {
                        "name": "send_mail",
                        "in": "query",
                        "description": "Whether to send an email with the password to the newly created member",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ],
                            "example": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Member registration successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MemberRegistration"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/{locale}/v1/member/login": {
            "post": {
                "tags": [
                    "Member",
                    "Authentication"
                ],
                "summary": "Log in a member",
                "description": "This endpoint allows a member to log in by providing their email and password.",
                "operationId": "loginMember",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "The locale (e.g. `en-us`)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "en-us"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The email of the member",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The password of the member",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MemberLoginSuccess"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "member_auth_token": []
                    }
                ]
            }
        },
        "/{locale}/v1/member/logout": {
            "post": {
                "tags": [
                    "Member",
                    "Authentication"
                ],
                "summary": "Log out a member",
                "description": "This endpoint revokes all access tokens for the authenticated member.",
                "operationId": "logoutMember",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "The locale (e.g., `en-us`)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "en-us"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Member logged out successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "member_auth_token": []
                    }
                ]
            }
        },
        "/{locale}/v1/member": {
            "get": {
                "tags": [
                    "Member",
                    "Authentication"
                ],
                "summary": "Get member data",
                "description": "This endpoint retrieves the data of the authenticated member.",
                "operationId": "getMember",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "The locale (e.g., `en-us`)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "en-us"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Member data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Member"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "member_auth_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "description": "The error message",
                        "type": "string",
                        "example": "The provided credentials are incorrect."
                    },
                    "code": {
                        "description": "The error code",
                        "type": "integer",
                        "example": 400
                    }
                },
                "type": "object"
            },
...
            "ValidationErrorResponse": {
                "properties": {
                    "message": {
                        "description": "The error message",
                        "type": "string",
                        "example": "The email must be a valid email address."
                    },
                    "errors": {
                        "description": "The validation errors",
                        "properties": {
                            "email": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "The email must be a valid email address."
                                ]
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "MemberLoginSuccess": {
                "required": [
                    "token"
                ],
                "properties": {
                    "token": {
                        "description": "The generated token for the member",
                        "type": "string",
                        "example": "21|5XDpYGurS96daVl7LSr34s03tqn37kfPzkbkqP6D"
                    }
                },
                "type": "object"
            },
            "MemberRegistration": {
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "email@example.com"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 64,
                        "example": "John Doe"
                    },
                    "password": {
                        "type": "string",
                        "maxLength": 48,
                        "example": "mypassword",
                        "nullable": true
                    },
                    "time_zone": {
                        "type": "string",
                        "example": "America/New_York",
                        "nullable": true
                    },
                    "locale": {
                        "type": "string",
                        "maxLength": 12,
                        "minLength": 5,
                        "example": "en_US",
                        "nullable": true
                    },
                    "currency": {
                        "type": "string",
                        "maxLength": 3,
                        "minLength": 3,
                        "example": "USD",
                        "nullable": true
                    },
                    "accepts_emails": {
                        "type": "integer",
                        "enum": [
                            0,
                            1
                        ],
                        "example": 1,
                        "nullable": true
                    },
                    "send_mail": {
                        "description": "If 1, send an email with the password to the newly created member.",
                        "type": "integer",
                        "enum": [
                            0,
                            1
                        ],
                        "example": 0,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Member": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "50510834128001805"
                    },
                    "unique_identifier": {
                        "type": "string",
                        "example": "700-857-223-945"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 64,
                        "example": "Member Name"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "member@example.com"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-06-20T09:12:01.000000Z"
                    },
                    "locale": {
                        "type": "string",
                        "maxLength": 12,
                        "minLength": 5,
                        "example": "en_US",
                        "nullable": true
                    },
                    "currency": {
                        "type": "string",
                        "maxLength": 3,
                        "minLength": 3,
                        "example": "USD",
                        "nullable": true
                    },
                    "time_zone": {
                        "type": "string",
                        "example": "America/New_York",
                        "nullable": true
                    },
                    "accepts_emails": {
                        "type": "integer",
                        "enum": [
                            0,
                            1
                        ],
                        "example": 0,
                        "nullable": true
                    },
                    "number_of_times_logged_in": {
                        "type": "integer",
                        "example": 6
                    },
                    "last_login_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-07-13T13:20:57.000000Z"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-06-20T09:12:01.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-07-13T13:20:57.000000Z"
                    },
                    "avatar": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "partner_auth_token": {
                "type": "http",
                "description": "Bearer {your_auth_token_here}",
                "name": "Authorization",
                "scheme": "bearer"
            },
            "member_auth_token": {
                "type": "http",
                "description": "Bearer {your_auth_token_here}",
                "name": "Authorization",
                "scheme": "bearer"
            }
        }
    }
}